+++ src/main/java/com/utime/burrowNest/admin/controller/AdminUserController.java
@@ -43,7 +43,7 @@ public class AdminUserController {
 	}
 	
 	@GetMapping("Profile.layer")
-    public String getUserProfile(ModelMap model, @RequestParam(name="userNo") int userNo) {
+    public String getUserProfile(ModelMap model, @RequestParam(name="userNo") long userNo) {
 		
 		model.addAttribute("item", userService.getUserFromNo(userNo));
 		model.addAttribute("roles", EJwtRole.values() );
