+++ src/main/java/com/utime/burrowNest/storage/service/StorageService.java
@@ -70,6 +70,6 @@ public interface StorageService {
 	 * @param groupNo
 	 * @return
 	 */
-	List<BnDirectory> getGroupStorageList(int groupNo);
+	List<BnDirectory> getGroupStorageList(long groupNo);
 
 }
