+++ src/main/java/com/utime/burrowNest/storage/vo/BnPathAccess.java
@@ -11,7 +11,7 @@ public class BnPathAccess {
 	/** 고유 번호 */
 	long no;
 	/** 그룹 번호 */
-	int groupNo;
+	long groupNo;
 	/** 엑세스 권한 */
 	int accessFlags;
