+++ src/main/java/com/utime/burrowNest/user/service/UserService.java
@@ -2,6 +2,6 @@ package com.utime.burrowNest.user.service;
 
 public interface UserService {
 
-	byte[] getThumbnail(int userNo);
+	byte[] getThumbnail(long userNo);
 	
 }
